capsules: spi: Rename Spi Controller file
Rename the SPI controller file to match the spi_peripherla.rs file, no
other chagnes to the file have been made.

Also update the README.
