+++ capsules/src/lib.rs
@@ -59,7 +59,7 @@ pub mod screen;
 pub mod sdcard;
 pub mod segger_rtt;
 pub mod si7021;
-pub mod spi;
+pub mod spi_controller;
 pub mod spi_peripheral;
 pub mod st7735;
 pub mod temperature;
